SC 110072: Organize components into evidence folder
