+++ src/applications/appeals/995/components/ConfirmationPageV2.jsx
@@ -6,9 +6,9 @@ import { selectProfile } from 'platform/user/selectors';
 
 // Content
 import { title995 } from '../content/title';
-import { EvidencePrivateContent } from './EvidencePrivateContent';
-import { EvidenceUploadContent } from './EvidenceUploadContent';
-import { EvidenceVaContent } from './EvidenceVaContent';
+import { PrivateContent } from './evidence/PrivateContent';
+import { UploadContent } from './evidence/UploadContent';
+import { VaContent } from './evidence/VaContent';
 import { content as notice5103Content } from '../content/notice5103';
 import { facilityTypeTitle, facilityTypeList } from '../content/facilityTypes';
 import { content as evidenceContent } from '../content/evidenceSummary';
@@ -194,7 +194,7 @@ export const ConfirmationPageV2 = () => {
       )}
 
       {vaEvidence.length ? (
-        <EvidenceVaContent
+        <VaContent
           list={vaEvidence}
           reviewMode
           showListOnly
@@ -203,7 +203,7 @@ export const ConfirmationPageV2 = () => {
       ) : null}
 
       {privateEvidence.length ? (
-        <EvidencePrivateContent
+        <PrivateContent
           list={privateEvidence}
           limitedConsent={data.limitedConsent}
           privacyAgreementAccepted={data.privacyAgreementAccepted}
@@ -215,7 +215,7 @@ export const ConfirmationPageV2 = () => {
       ) : null}
 
       {otherEvidence.length ? (
-        <EvidenceUploadContent
+        <UploadContent
           list={otherEvidence}
           reviewMode
           showListOnly
