+++ src/applications/appeals/995/components/evidence/PageNavigation.jsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import { Link } from 'react-router';
 import FormNavButtons from 'platform/forms-system/src/js/components/FormNavButtons';
 
-export const EvidencePageNavigation = ({ path, content, handlers }) => (
+export const PageNavigation = ({ path, content, handlers }) => (
   <>
     <div className="vads-u-margin-top--2">
       <Link
@@ -26,7 +26,7 @@ export const EvidencePageNavigation = ({ path, content, handlers }) => (
   </>
 );
 
-EvidencePageNavigation.propTypes = {
+PageNavigation.propTypes = {
   content: PropTypes.shape({
     addAnotherLink: PropTypes.string,
     contentAfterButtons: PropTypes.element,
