+++ src/applications/appeals/995/components/evidence/PrivateRecordsRequest.jsx
@@ -4,22 +4,26 @@ import { VaRadio } from '@department-of-veterans-affairs/component-library/dist/
 import FormNavButtons from 'platform/forms-system/src/js/components/FormNavButtons';
 import recordEvent from 'platform/monitoring/record-event';
 
-import { EVIDENCE_VA_PATH, EVIDENCE_VA, EVIDENCE_PRIVATE } from '../constants';
+import {
+  EVIDENCE_VA_PATH,
+  EVIDENCE_VA,
+  EVIDENCE_PRIVATE,
+} from '../../constants';
 
 import {
   privateRecordsRequestTitle,
   privateRecordsRequestInfo,
   privateRecordsRadioDescription,
-} from '../content/evidencePrivateRecordsRequest';
+} from '../../content/evidencePrivateRecordsRequest';
 
-import { customPageProps995 } from '../../shared/props';
-import errorMessages from '../../shared/content/errorMessages';
-import { focusFirstError } from '../../shared/utils/focus';
+import { customPageProps995 } from '../../../shared/props';
+import errorMessages from '../../../shared/content/errorMessages';
+import { focusFirstError } from '../../../shared/utils/focus';
 
 /**
  * This page is needed to make the back button on this page to to the last
  */
-const EvidencePrivateRequest = ({
+const PrivateRecordsRequest = ({
   data = {},
   goBack,
   goForward,
@@ -103,6 +107,6 @@ const EvidencePrivateRequest = ({
   );
 };
 
-EvidencePrivateRequest.propTypes = customPageProps995;
+PrivateRecordsRequest.propTypes = customPageProps995;
 
-export default EvidencePrivateRequest;
+export default PrivateRecordsRequest;
