+++ src/applications/appeals/995/tests/components/evidence/PrivateContent.unit.spec.jsx
@@ -7,16 +7,16 @@ import {
   EVIDENCE_PRIVATE_PATH,
   EVIDENCE_LIMITATION_PATH,
   LIMITATION_KEY,
-} from '../../constants';
-import { content } from '../../content/evidenceSummary';
-import { EvidencePrivateContent } from '../../components/EvidencePrivateContent';
-import { records } from '../data/evidence-records';
+} from '../../../constants';
+import { content } from '../../../content/evidenceSummary';
+import { PrivateContent } from '../../../components/evidence/PrivateContent';
+import { records } from '../../data/evidence-records';
 
 describe('buildPrivateContent', () => {
   it('should render editable private content', () => {
     const privateEvidence = records().providerFacility;
     const { container } = render(
-      <EvidencePrivateContent
+      <PrivateContent
         list={privateEvidence}
         limitedConsent="test"
         privacyAgreementAccepted
@@ -43,7 +43,7 @@ describe('buildPrivateContent', () => {
   it('should render nothing when no data is passed in', () => {
     const { container } = render(
       <div>
-        <EvidencePrivateContent testing />
+        <PrivateContent testing />
       </div>,
     );
 
@@ -53,11 +53,7 @@ describe('buildPrivateContent', () => {
   it('should not render limited consent section remove button', () => {
     const privateEvidence = records().providerFacility;
     const { container } = render(
-      <EvidencePrivateContent
-        list={privateEvidence}
-        limitedConsent=""
-        testing
-      />,
+      <PrivateContent list={privateEvidence} limitedConsent="" testing />,
     );
     expect($$('.private-facility', container).length).to.eq(2);
     expect($$('.private-limitation', container).length).to.eq(1);
@@ -68,7 +64,7 @@ describe('buildPrivateContent', () => {
   it('should render review-only private content', () => {
     const privateEvidence = records().providerFacility;
     const { container } = render(
-      <EvidencePrivateContent
+      <PrivateContent
         list={privateEvidence}
         limitedConsent="test"
         reviewMode
@@ -90,7 +86,7 @@ describe('buildPrivateContent', () => {
   it('should render list only for confirmation page content', () => {
     const privateEvidence = records().providerFacility;
     const { container } = render(
-      <EvidencePrivateContent
+      <PrivateContent
         list={privateEvidence}
         limitedConsent="test"
         reviewMode
@@ -112,7 +108,7 @@ describe('buildPrivateContent', () => {
 
   it('should show missing facility message', () => {
     const { container } = render(
-      <EvidencePrivateContent list={['']} limitedConsent="" testing />,
+      <PrivateContent list={['']} limitedConsent="" testing />,
     );
 
     const ul = $('ul', container);
@@ -125,11 +121,7 @@ describe('buildPrivateContent', () => {
   it('should show missing issues message', () => {
     const privateEvidence = records({ emptyIssue: true }).providerFacility;
     const { container } = render(
-      <EvidencePrivateContent
-        list={privateEvidence}
-        limitedConsent=""
-        testing
-      />,
+      <PrivateContent list={privateEvidence} limitedConsent="" testing />,
     );
 
     const li = $$('li', container);
@@ -140,7 +132,7 @@ describe('buildPrivateContent', () => {
   it('should have edit links pointing to the appropriate private indexed page or limitation page', () => {
     const privateEvidence = records().providerFacility;
     const { container } = render(
-      <EvidencePrivateContent list={privateEvidence} testing />,
+      <PrivateContent list={privateEvidence} testing />,
     );
 
     const links = $$('.edit-item', container);
@@ -160,11 +152,7 @@ describe('buildPrivateContent', () => {
     const privateEvidence = records().providerFacility;
     const handlers = { showModal: removeSpy };
     const { container } = render(
-      <EvidencePrivateContent
-        list={privateEvidence}
-        handlers={handlers}
-        testing
-      />,
+      <PrivateContent list={privateEvidence} handlers={handlers} testing />,
     );
 
     const buttons = $$('.remove-item', container);
@@ -187,7 +175,7 @@ describe('buildPrivateContent', () => {
     const privateEvidence = records().providerFacility;
     const handlers = { showModal: removeSpy };
     const { container } = render(
-      <EvidencePrivateContent
+      <PrivateContent
         list={privateEvidence}
         limitedConsent="test"
         handlers={handlers}
