+++ src/applications/appeals/995/tests/components/evidence/Summary.unit.spec.jsx
@@ -5,8 +5,8 @@ import sinon from 'sinon';
 
 import { $, $$ } from 'platform/forms-system/src/js/utilities/ui';
 
-import EvidenceSummary from '../../components/EvidenceSummary';
-import { content } from '../../content/evidenceSummary';
+import Summary from '../../../components/evidence/Summary';
+import { content } from '../../../content/evidenceSummary';
 import {
   EVIDENCE_PRIVATE,
   EVIDENCE_VA,
@@ -18,7 +18,7 @@ import {
   EVIDENCE_UPLOAD_PATH,
   EVIDENCE_LIMIT,
   EVIDENCE_PRIVATE_AUTHORIZATION,
-} from '../../constants';
+} from '../../../constants';
 
 const providerFacilityAddress = {
   country: 'USA',
@@ -90,7 +90,7 @@ const setupSummary = ({
 } = {}) =>
   render(
     <div>
-      <EvidenceSummary
+      <Summary
         data={{
           [EVIDENCE_VA]: vaMR,
           [EVIDENCE_PRIVATE]: privateMR,
@@ -114,7 +114,7 @@ const setupSummary = ({
     </div>,
   );
 
-describe('<EvidenceSummary>', () => {
+describe('<Summary>', () => {
   it('should render', () => {
     const { container } = setupSummary({ limit: 'Pizza addiction' });
