+++ src/applications/appeals/995/tests/components/evidence/VaRecords.unit.spec.jsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import { expect } from 'chai';
 import { render, fireEvent, waitFor } from '@testing-library/react';
+import userEvent from '@testing-library/user-event';
 import sinon from 'sinon';
 
 import {
@@ -10,17 +11,23 @@ import {
 
 import * as focusUtils from '~/platform/utilities/ui/focus';
 
-import EvidencePrivateRecords from '../../components/EvidencePrivateRecords';
+import VaRecords from '../../../components/evidence/VaRecords';
 import {
   errorMessages,
-  EVIDENCE_PRIVATE_PATH,
+  EVIDENCE_VA,
+  EVIDENCE_VA_PATH,
   NO_ISSUES_SELECTED,
-} from '../../constants';
+} from '../../../constants';
 
-import { clickAddAnother, clickBack, clickContinue } from './helpers';
-import { parseDateWithOffset } from '../../../shared/utils/dates';
-import { SELECTED, MAX_YEARS_PAST } from '../../../shared/constants';
-import sharedErrorMessages from '../../../shared/content/errorMessages';
+import { clickAddAnother, clickBack, clickContinue } from '../helpers';
+
+import { parseDateWithOffset } from '../../../../shared/utils/dates';
+import {
+  MAX_LENGTH,
+  MAX_YEARS_PAST,
+  SELECTED,
+} from '../../../../shared/constants';
+import sharedErrorMessage from '../../../../shared/content/errorMessages';
 
 /*
 | Data     | Forward     | Back               | Add another      |
@@ -29,7 +36,15 @@ import sharedErrorMessages from '../../../shared/content/errorMessages';
 | Empty    | Focus error | Prev page & remove | Focus error      |
 | Partial  | Focus error | Modal & Prev page  | Focus error      |
  */
-describe('<EvidencePrivateRecords>', () => {
+describe('<VaRecords>', () => {
+  let focusElementSpy;
+  beforeEach(() => {
+    focusElementSpy = sinon.stub(focusUtils, 'focusElement');
+  });
+  afterEach(() => {
+    focusElementSpy.restore();
+  });
+
   const validDate = parseDateWithOffset({ months: -2 });
   const mockData = {
     contestedIssues: [
@@ -46,25 +61,27 @@ describe('<EvidencePrivateRecords>', () => {
       { issue: 'test 2', decisionDate: validDate, [SELECTED]: true },
     ],
   };
-  const mockAddress = {
-    country: 'USA',
-    street: '123 Main',
-    street2: '',
-    city: 'City',
-    state: 'CA',
-    postalCode: '90210',
-  };
-  const mockFacility = {
-    providerFacilityName: 'Location 1',
-    providerFacilityAddress: mockAddress,
+  const mockLocation = {
+    locationAndName: 'Location 1',
     issues: ['test 1'],
-    treatmentDateRange: { from: '2001-01-01', to: '2011-01-01' },
+    evidenceDates: { from: '2001-01-01', to: '2011-01-01' },
   };
-  const mockFacility2 = {
-    providerFacilityName: 'Location 2',
-    providerFacilityAddress: mockAddress,
+  const mockLocation2 = {
+    locationAndName: 'Location 2',
     issues: ['test 2'],
-    treatmentDateRange: { from: '2002-02-02', to: '2012-02-02' },
+    evidenceDates: { from: '2002-02-02', to: '2012-02-02' },
+  };
+
+  const mockLocationNew = {
+    locationAndName: 'Location 3',
+    issues: ['test 1'],
+    treatmentDate: '2005-05',
+  };
+  const mockLocationNew2 = {
+    locationAndName: 'Location 4',
+    issues: ['test 2'],
+    treatmentDate: '',
+    noDate: true,
   };
 
   const setup = ({
@@ -76,7 +93,7 @@ describe('<EvidencePrivateRecords>', () => {
     setFormData = () => {},
   } = {}) => (
     <div>
-      <EvidencePrivateRecords
+      <VaRecords
         testingIndex={index}
         data={data}
         goBack={goBack}
@@ -91,43 +108,26 @@ describe('<EvidencePrivateRecords>', () => {
 
   const getErrorElements = container =>
     $$(
-      [
-        'va-text-input[error]',
-        'va-select[error]',
-        'va-checkbox-group[error]',
-        'va-date[error]',
-        'va-memorable-date[error]',
-      ].join(','),
+      'va-text-input[error], va-checkbox-group[error], va-date[error], va-memorable-date[error]',
       container,
     );
 
   it('should render', () => {
     const { container } = render(setup());
+    expect($('h3', container)).to.exist;
     expect($('va-modal', container)).to.exist;
-    expect($$('va-text-input', container).length).to.eq(5);
-    expect($$('va-select', container).length).to.eq(2);
+    expect($('va-text-input', container)).to.exist;
     expect($('va-checkbox-group', container)).to.exist;
     expect($$('va-checkbox', container).length).to.eq(2);
     expect($$('va-memorable-date', container).length).to.eq(2);
     expect($('.vads-c-action-link--green', container)).to.exist;
     // check Datadog classes
     expect(
-      $$('va-checkbox.dd-privacy-hidden[data-dd-action-name]', container)
-        .length,
+      $$('.dd-privacy-hidden[data-dd-action-name]', container).length,
     ).to.eq(2);
   });
 
-  it('should render with no data', () => {
-    const { container } = render(setup({ data: null }));
-    expect($$('va-text-input', container).length).to.eq(5);
-    expect($$('va-select', container).length).to.eq(2);
-    expect($('va-checkbox-group', container)).to.exist;
-    expect($$('va-checkbox', container).length).to.eq(0);
-    expect($$('va-memorable-date', container).length).to.eq(2);
-    expect($('.vads-c-action-link--green', container)).to.exist;
-  });
-
-  it('should update facility name', async () => {
+  it('should update location name', async () => {
     const setDataSpy = sinon.spy();
     const page = setup({ setFormData: setDataSpy });
     const { container } = render(page);
@@ -137,18 +137,12 @@ describe('<EvidencePrivateRecords>', () => {
     fireEvent.input(input, { target: { name: 'name' } });
 
     expect(setDataSpy.called).to.be.true;
-    expect(setDataSpy.args[0][0].providerFacility[0]).to.deep.equal({
-      providerFacilityName: input.value,
-      providerFacilityAddress: {
-        country: 'USA',
-        street: '',
-        street2: '',
-        city: '',
-        state: '',
-        postalCode: '',
-      },
+    expect(setDataSpy.args[0][0].locations[0]).to.deep.equal({
+      locationAndName: input.value,
       issues: [],
-      treatmentDateRange: { from: '', to: '' },
+      evidenceDates: { from: '', to: '' },
+      noDate: undefined,
+      treatmentDate: '',
     });
   });
 
@@ -156,7 +150,7 @@ describe('<EvidencePrivateRecords>', () => {
     const setDataSpy = sinon.spy();
     const page = setup({
       setFormData: setDataSpy,
-      data: { ...mockData, providerFacility: [mockFacility] },
+      data: { ...mockData, [EVIDENCE_VA]: true, locations: [mockLocation] },
     });
     const { container } = render(page);
 
@@ -166,9 +160,11 @@ describe('<EvidencePrivateRecords>', () => {
     });
 
     expect(setDataSpy.called).to.be.true;
-    expect(setDataSpy.args[0][0].providerFacility[0]).to.deep.equal({
-      ...mockFacility,
+    expect(setDataSpy.args[0][0].locations[0]).to.deep.equal({
+      ...mockLocation,
       issues: ['test 1', 'test 2'],
+      noDate: undefined,
+      treatmentDate: undefined,
     });
   });
 
@@ -176,7 +172,7 @@ describe('<EvidencePrivateRecords>', () => {
     const setDataSpy = sinon.spy();
     const page = setup({
       setFormData: setDataSpy,
-      data: { ...mockData, providerFacility: [mockFacility] },
+      data: { ...mockData, [EVIDENCE_VA]: true, locations: [mockLocation] },
     });
     const { container } = render(page);
 
@@ -186,17 +182,23 @@ describe('<EvidencePrivateRecords>', () => {
     });
 
     expect(setDataSpy.called).to.be.true;
-    expect(setDataSpy.args[0][0].providerFacility[0]).to.deep.equal({
-      ...mockFacility,
+    expect(setDataSpy.args[0][0].locations[0]).to.deep.equal({
+      ...mockLocation,
       issues: [],
+      noDate: undefined,
+      treatmentDate: undefined,
     });
   });
 
   // *** VALID DATA ***
   describe('valid data navigation', () => {
-    it('should navigate forward to limitation page with valid data', async () => {
+    it('should navigate forward to VA private request page with valid data', async () => {
       const goSpy = sinon.spy();
-      const data = { ...mockData, providerFacility: [mockFacility] };
+      const data = {
+        ...mockData,
+        [EVIDENCE_VA]: true,
+        locations: [mockLocation],
+      };
       const page = setup({
         index: 0,
         goForward: goSpy,
@@ -206,11 +208,14 @@ describe('<EvidencePrivateRecords>', () => {
 
       // continue
       clickContinue(container);
-      await waitFor(() => expect(goSpy.calledWith(data)).to.be.true);
+      await waitFor(() => {
+        expect($('va-modal[visible="false"]', container)).to.exist;
+        expect(goSpy.calledWith(data)).to.be.true;
+      });
     });
-    it('should navigate back to private records request page with valid data', async () => {
+    it('should navigate back to VA records request page with valid data', async () => {
       const goSpy = sinon.spy();
-      const data = { ...mockData, providerFacility: [mockFacility] };
+      const data = { ...mockData, locations: [mockLocation] };
       const index = 0;
       const page = setup({
         index,
@@ -221,14 +226,19 @@ describe('<EvidencePrivateRecords>', () => {
 
       // back
       clickBack(container);
-      // passing a negative index is okay, we're leaving the indexed pages
-      await waitFor(() => expect(goSpy.calledWith(index - 1)).to.be.true);
+
+      await waitFor(() => {
+        expect($('va-modal[visible="false"]', container)).to.exist;
+        // passing a negative index is okay, we're leaving the indexed pages
+        expect(goSpy.calledWith(index - 1)).to.be.true;
+      });
     });
-    it('should navigate from zero index to a new empty facility page, of index 1, with valid data', async () => {
+    it('should navigate from zero index to a new empty location page, of index 1, with valid data', async () => {
       const goSpy = sinon.spy();
       const data = {
         ...mockData,
-        providerFacility: [mockFacility, {}, mockFacility2],
+        [EVIDENCE_VA]: true,
+        locations: [mockLocation, {}, mockLocation2],
       };
       const index = 0;
       const page = setup({
@@ -239,18 +249,19 @@ describe('<EvidencePrivateRecords>', () => {
       const { container } = render(page);
 
       // add
-      await clickAddAnother(container);
+      clickAddAnother(container);
 
-      waitFor(() => {
+      await waitFor(() => {
         expect($('va-modal[visible="false"]', container)).to.exist;
-        expect(goSpy.calledWith(`/${EVIDENCE_PRIVATE_PATH}?index=${index + 1}`))
-          .to.be.true;
+        expect(goSpy.calledWith(`/${EVIDENCE_VA_PATH}?index=${index + 1}`)).to
+          .be.true;
       });
     });
+
     it('should navigate from zero index, with valid data, to next index when inserting another entry', async () => {
       const goSpy = sinon.spy();
-      const providerFacility = [mockFacility, mockFacility2, {}];
-      const data = { ...mockData, providerFacility };
+      const locations = [mockLocation, mockLocation2, {}];
+      const data = { ...mockData, [EVIDENCE_VA]: true, locations };
       const index = 0;
       const page = setup({
         index,
@@ -260,68 +271,182 @@ describe('<EvidencePrivateRecords>', () => {
       const { container } = render(page);
 
       // add
-      await clickAddAnother(container);
+      clickAddAnother(container);
 
-      waitFor(async () => {
+      await waitFor(() => {
         expect($('va-modal[visible="false"]', container)).to.exist;
-        expect(goSpy.calledWith(`/${EVIDENCE_PRIVATE_PATH}?index=${index + 1}`))
-          .to.be.true;
+        expect(goSpy.calledWith(`/${EVIDENCE_VA_PATH}?index=${index + 1}`)).to
+          .be.true;
+      });
+    });
+
+    /* New SC form content */
+    const showScNewForm = true;
+    it('should navigate forward to VA private request page with YYYY-MM date & valid data', async () => {
+      const goSpy = sinon.spy();
+      const data = {
+        ...mockData,
+        showScNewForm,
+        [EVIDENCE_VA]: true,
+        locations: [mockLocationNew],
+      };
+      const page = setup({
+        index: 0,
+        goForward: goSpy,
+        data,
+      });
+      const { container } = render(page);
+
+      // continue
+      clickContinue(container);
+      await waitFor(() => {
+        expect($('va-modal[visible="false"]', container)).to.exist;
+        expect(goSpy.calledWith(data)).to.be.true;
+      });
+    });
+    it('should navigate forward to VA private request page with no date checked & valid data', async () => {
+      const goSpy = sinon.spy();
+      const data = {
+        ...mockData,
+        showScNewForm,
+        [EVIDENCE_VA]: true,
+        locations: [mockLocationNew2],
+      };
+      const page = setup({
+        index: 0,
+        goForward: goSpy,
+        data,
+      });
+      const { container } = render(page);
+
+      // continue
+      clickContinue(container);
+      await waitFor(() => {
+        expect($('va-modal[visible="false"]', container)).to.exist;
+        expect(goSpy.calledWith(data)).to.be.true;
+      });
+    });
+    it('should navigate back to VA records request page with YYYY-MM date & valid data', async () => {
+      const goSpy = sinon.spy();
+      const data = {
+        ...mockData,
+        showScNewForm,
+        [EVIDENCE_VA]: true,
+        locations: [mockLocationNew],
+      };
+      const index = 0;
+      const page = setup({
+        index,
+        goBack: goSpy,
+        data,
+      });
+      const { container } = render(page);
+
+      // back
+      clickBack(container);
+
+      await waitFor(() => {
+        expect($('va-modal[visible="false"]', container)).to.exist;
+        // passing a negative index is okay, we're leaving the indexed pages
+        expect(goSpy.calledWith(index - 1)).to.be.true;
+      });
+    });
+    it('should navigate from zero index to a new empty location page, of index 1, with YYYY-MM date & valid data', async () => {
+      const goSpy = sinon.spy();
+      const data = {
+        ...mockData,
+        showScNewForm,
+        [EVIDENCE_VA]: true,
+        locations: [mockLocationNew, {}, mockLocationNew2],
+      };
+      const index = 0;
+      const page = setup({
+        index,
+        goToPath: goSpy,
+        data,
+      });
+      const { container } = render(page);
+
+      // add
+      clickAddAnother(container);
+
+      await waitFor(() => {
+        expect($('va-modal[visible="false"]', container)).to.exist;
+        expect(goSpy.calledWith(`/${EVIDENCE_VA_PATH}?index=${index + 1}`)).to
+          .be.true;
+      });
+    });
+
+    it('should navigate from zero index, with YYYY-MM date & valid data, to next index when inserting another entry', async () => {
+      const goSpy = sinon.spy();
+      const locations = [mockLocationNew, mockLocationNew2, {}];
+      const data = {
+        ...mockData,
+        showScNewForm,
+        [EVIDENCE_VA]: true,
+        locations,
+      };
+      const index = 0;
+      const page = setup({
+        index,
+        goToPath: goSpy,
+        data,
+      });
+      const { container } = render(page);
+
+      // add
+      clickAddAnother(container);
+
+      await waitFor(() => {
+        expect($('va-modal[visible="false"]', container)).to.exist;
+        expect(goSpy.calledWith(`/${EVIDENCE_VA_PATH}?index=${index + 1}`)).to
+          .be.true;
       });
     });
   });
 
   // *** EMPTY PAGE ***
   describe('empty page navigation', () => {
-    const getAndTestAllErrors = async (container, options = {}) => {
+    const getAndTestAllErrors = async container => {
       const errors = errorMessages.evidence;
       const errorEls = await getErrorElements(container);
-      [
-        errors.facilityMissing,
-        options.ignoreCountry ? null : sharedErrorMessages.country, // default set to USA
-        sharedErrorMessages.street,
-        sharedErrorMessages.city,
-        sharedErrorMessages.state,
-        sharedErrorMessages.postal,
-        errors.issuesMissing,
-        errors.blankDate,
-        errors.blankDate,
-      ]
-        .filter(Boolean)
-        .forEach((error, index) => {
-          expect(errorEls[index]?.error).to.eq(error);
-          if (error === errors.blankDate) {
-            expect(errorEls[index].invalidMonth).to.be.true;
-            expect(errorEls[index].invalidDay).to.be.true;
-            expect(errorEls[index].invalidYear).to.be.true;
-          }
-        });
+      expect(errorEls[0].error).to.eq(errors.locationMissing);
+      expect(errorEls[1].error).to.eq(errors.issuesMissing);
+
+      expect(errorEls[2].error).to.eq(errorMessages.evidence.blankDate);
+      expect(errorEls[2].invalidMonth).to.be.true;
+      expect(errorEls[2].invalidDay).to.be.true;
+      expect(errorEls[2].invalidYear).to.be.true;
+
+      expect(errorEls[3].error).to.eq(errorMessages.evidence.blankDate);
+      expect(errorEls[3].invalidMonth).to.be.true;
+      expect(errorEls[3].invalidDay).to.be.true;
+      expect(errorEls[3].invalidYear).to.be.true;
     };
 
-    it('should show & focus error messages after going forward on an empty first page', async () => {
+    it('should show & focus on error messages after going forward on an empty first page', async () => {
       const goSpy = sinon.spy();
       const index = 0;
-      const page = setup({
-        index,
-        goForward: goSpy,
-        goToPath: goSpy,
-      });
+      const page = setup({ index, goForward: goSpy });
       const { container } = render(page);
 
-      // continue
-      await clickContinue(container);
+      // forward
+      clickContinue(container);
 
-      waitFor(async () => {
+      await waitFor(() => {
         expect($('va-modal[visible="false"]', container)).to.exist;
         expect(goSpy.called).to.be.false;
-        getAndTestAllErrors(container, { ignoreCountry: true });
+      }).then(() => {
+        getAndTestAllErrors(container);
       });
     });
-    it('should show & focus error messages after going forward on an empty second page', async () => {
+    it('should show & focus on error messages after going forward on an empty second page', async () => {
       const goSpy = sinon.spy();
       const index = 1;
       const data = {
         ...mockData,
-        providerFacility: [mockFacility, {}, mockFacility2],
+        [EVIDENCE_VA]: true,
+        locations: [mockLocation, {}, mockLocation2],
       };
       const page = setup({
         index,
@@ -337,13 +462,14 @@ describe('<EvidencePrivateRecords>', () => {
       await waitFor(() => {
         expect($('va-modal[visible="false"]', container)).to.exist;
         expect(goSpy.called).to.be.false;
+      }).then(() => {
         getAndTestAllErrors(container);
       });
     });
     it('should go back on an empty page on first entry', async () => {
       const goSpy = sinon.spy();
       const index = 0;
-      const page = setup({ index, goBack: goSpy, goToPath: goSpy });
+      const page = setup({ index, goBack: goSpy });
       const { container } = render(page);
 
       // back
@@ -359,7 +485,11 @@ describe('<EvidencePrivateRecords>', () => {
     it('should go back and remove empty page', async () => {
       const goSpy = sinon.spy();
       const setDataSpy = sinon.spy();
-      const data = { ...mockData, providerFacility: [mockFacility, {}] };
+      const data = {
+        ...mockData,
+        [EVIDENCE_VA]: true,
+        locations: [mockLocation, {}],
+      };
       const page = setup({
         index: 1,
         goToPath: goSpy,
@@ -375,7 +505,7 @@ describe('<EvidencePrivateRecords>', () => {
         expect($('va-modal[visible="false"]', container)).to.exist;
         expect(goSpy.called).to.be.true;
         expect(setDataSpy.called).to.be.true;
-        expect(setDataSpy.lastCall.args[0].providerFacility.length).to.eq(1);
+        expect(setDataSpy.lastCall.args[0].locations.length).to.eq(1);
       });
     });
     it('should show & focus on error messages after adding new location on an empty second page', async () => {
@@ -383,7 +513,8 @@ describe('<EvidencePrivateRecords>', () => {
       const index = 1;
       const data = {
         ...mockData,
-        providerFacility: [mockFacility, {}, mockFacility2],
+        [EVIDENCE_VA]: true,
+        locations: [mockLocation, {}, mockLocation2],
       };
       const page = setup({
         index,
@@ -399,9 +530,11 @@ describe('<EvidencePrivateRecords>', () => {
       await waitFor(() => {
         expect($('va-modal[visible="false"]', container)).to.exist;
         expect(goSpy.called).to.be.false;
+      }).then(() => {
         getAndTestAllErrors(container);
       });
     });
+
     it('should cancel navigation', async () => {
       const goSpy = sinon.spy();
       const index = 0;
@@ -414,45 +547,34 @@ describe('<EvidencePrivateRecords>', () => {
       const event = new CustomEvent('closeEvent');
       await $('va-modal', container).__events.closeEvent(event);
 
-      waitFor(() => {
+      await waitFor(() => {
         expect($('va-modal[visible="false"]', container)).to.exist;
       });
     });
   });
 
   describe('partial/invalid data navigation', () => {
-    let focusElementSpy;
-    beforeEach(() => {
-      focusElementSpy = sinon.stub(focusUtils, 'focusElement');
-    });
-    afterEach(() => {
-      focusElementSpy.restore();
-    });
-
-    const testAndCloseModal = (container, total, event) =>
+    const testAndCloseModal = async (container, event) => {
       // modal visible
       waitFor(() => {
         expect($('va-modal[visible="true"]', container)).to.exist;
       })
         .then(() => {
-          expect(getErrorElements(container).length).to.eq(total);
           // close modal
           $('va-modal').__events[event]();
         })
         .finally(() => {
           expect($('va-modal[visible="false"]', container)).to.exist;
         });
+    };
 
     it('should not navigate, but will show errors after continuing', async () => {
       const goSpy = sinon.spy();
       const index = 1;
       const data = {
         ...mockData,
-        providerFacility: [
-          mockFacility,
-          { providerFacilityName: 'foo' },
-          mockFacility2,
-        ],
+        [EVIDENCE_VA]: true,
+        locations: [mockLocation, { locationAndName: 'foo' }, mockLocation2],
       };
       const page = setup({
         index,
@@ -463,14 +585,12 @@ describe('<EvidencePrivateRecords>', () => {
       const { container } = render(page);
 
       // continue
-      clickContinue(container);
+      await clickContinue(container);
 
       waitFor(() => {
         expect(goSpy.called).to.be.false;
-        expect(focusElementSpy.called).to.be.true;
+        expect(getErrorElements(container).length).to.eq(3);
         expect(focusElementSpy.args[0][0]).to.eq('[role="alert"]');
-      }).then(() => {
-        expect(getErrorElements(container).length).to.eq(8);
       });
     });
 
@@ -481,11 +601,8 @@ describe('<EvidencePrivateRecords>', () => {
       const index = 1;
       const data = {
         ...mockData,
-        providerFacility: [
-          mockFacility,
-          { providerFacilityName: 'foo' },
-          mockFacility2,
-        ],
+        [EVIDENCE_VA]: true,
+        locations: [mockLocation, { locationAndName: 'foo' }, mockLocation2],
       };
       const page = setup({
         index,
@@ -498,13 +615,18 @@ describe('<EvidencePrivateRecords>', () => {
 
       // back
       await clickBack(container);
-      // Click no
-      testAndCloseModal(container, 8, 'secondaryButtonClick').then(() => {
+
+      // This check is super-flaky in CI
+      waitFor(() => {
+        expect(getErrorElements(container).length).to.eq(3);
+      });
+
+      testAndCloseModal(container, 'secondaryButtonClick').then(() => {
         expect(setDataSpy.called).to.be.true;
-        expect(setDataSpy.lastCall.args[0].providerFacility.length).to.eq(2);
+        expect(setDataSpy.lastCall.args[0].locations.length).to.eq(2);
         expect(goSpy.called).to.be.true;
-        expect(goSpy.calledWith(`/${EVIDENCE_PRIVATE_PATH}?index=${index - 1}`))
-          .to.be.true;
+        expect(goSpy.calledWith(`/${EVIDENCE_VA_PATH}?index=${index - 1}`)).to
+          .be.true;
       });
     });
 
@@ -518,23 +640,26 @@ describe('<EvidencePrivateRecords>', () => {
         setFormData: setDataSpy,
         data: {
           ...mockData,
-          providerFacility: [
-            mockFacility,
-            mockFacility2,
-            { providerFacilityName: 'foo' },
-          ],
+          [EVIDENCE_VA]: true,
+          locations: [mockLocation, mockLocation2, { locationAndName: 'foo' }],
         },
       });
       const { container } = render(page);
 
       // back
       await clickBack(container);
-      // Click yes to keep partial entry
-      testAndCloseModal(container, 8, 'primaryButtonClick').then(() => {
+
+      // This check is super-flaky in CI
+      waitFor(() => {
+        expect(getErrorElements(container).length).to.eq(3);
+      });
+
+      // keep partial entry
+      testAndCloseModal(container, 'primaryButtonClick').then(() => {
         expect(setDataSpy.called).to.be.false; // no data change
         expect(goSpy.called).to.be.true;
-        expect(goSpy.calledWith(`/${EVIDENCE_PRIVATE_PATH}?index=${index - 1}`))
-          .to.be.true;
+        expect(goSpy.calledWith(`/${EVIDENCE_VA_PATH}?index=${index - 1}`)).to
+          .be.true;
       });
     });
 
@@ -547,20 +672,19 @@ describe('<EvidencePrivateRecords>', () => {
         goToPath: goSpy,
         data: {
           ...mockData,
-          providerFacility: [mockFacility, { lproviderFacilityName: 'foo' }],
+          [EVIDENCE_VA]: true,
+          locations: [mockLocation, { locationAndName: 'foo' }],
         },
       });
       const { container } = render(page);
 
       // add
-      clickAddAnother(container);
+      await clickAddAnother(container);
 
-      waitFor(async () => {
+      waitFor(() => {
         expect(goSpy.called).to.be.false;
-        expect(focusElementSpy.called).to.be.true;
+        expect(getErrorElements(container).length).to.eq(3);
         expect(focusElementSpy.args[0][0]).to.eq('[role="alert"]');
-      }).then(() => {
-        expect(getErrorElements(container).length).to.eq(9);
       });
     });
   });
@@ -569,20 +693,44 @@ describe('<EvidencePrivateRecords>', () => {
     const fromBlurEvent = new CustomEvent('blur', { detail: 'from' });
     const toBlurEvent = new CustomEvent('blur', { detail: 'to' });
 
-    it('should show error when start treatment date is in the future', async () => {
-      const from = parseDateWithOffset({ years: 1 });
+    // *** OTHER ERRORS ***
+    it('should show error when location name is too long', async () => {
+      const name = 'abcdef '.repeat(
+        MAX_LENGTH.SC_EVIDENCE_LOCATION_AND_NAME / 6,
+      );
       const data = {
         ...mockData,
-        providerFacility: [{ treatmentDateRange: { from } }],
+        [EVIDENCE_VA]: true,
+        locations: [{ locationAndName: name }],
       };
       const page = setup({ index: 0, data });
       const { container } = render(page);
 
-      const dateFrom = await $('va-memorable-date', container);
+      const input = await $('va-text-input', container);
+      await fireEvent.blur(input);
+
+      waitFor(() => {
+        expect(input.error).to.contain(
+          errorMessages.evidence.locationMaxLength,
+        );
+      });
+    });
+
+    it('should show error when start treatment date is in the future', async () => {
+      const from = parseDateWithOffset({ years: 1 });
+      const data = {
+        ...mockData,
+        [EVIDENCE_VA]: true,
+        locations: [{ evidenceDates: { from } }],
+      };
+      const page = setup({ index: 0, data });
+      const { container } = render(page);
+
       // blur date from input
       await $('va-memorable-date').__events.dateBlur(fromBlurEvent);
 
       waitFor(async () => {
+        const dateFrom = await $('va-memorable-date', container);
         expect(dateFrom.error).to.contain(errorMessages.evidence.pastDate);
         expect(dateFrom.invalidMonth).to.be.false;
         expect(dateFrom.invalidDay).to.be.false;
@@ -594,7 +742,8 @@ describe('<EvidencePrivateRecords>', () => {
       const to = parseDateWithOffset({ years: 1 });
       const data = {
         ...mockData,
-        providerFacility: [{ treatmentDateRange: { to } }],
+        [EVIDENCE_VA]: true,
+        locations: [{ evidenceDates: { to } }],
       };
       const page = setup({ index: 0, data });
       const { container } = render(page);
@@ -611,11 +760,12 @@ describe('<EvidencePrivateRecords>', () => {
       });
     });
 
-    it('should show an error when the start treatment date is too far in the past', async () => {
+    it('should show an error when the start treament date is too far in the past', async () => {
       const from = parseDateWithOffset({ years: -(MAX_YEARS_PAST + 1) });
       const data = {
         ...mockData,
-        providerFacility: [{ treatmentDateRange: { from } }],
+        [EVIDENCE_VA]: true,
+        locations: [{ evidenceDates: { from } }],
       };
       const page = setup({ index: 0, data });
       const { container } = render(page);
@@ -636,7 +786,8 @@ describe('<EvidencePrivateRecords>', () => {
       const to = parseDateWithOffset({ years: -(MAX_YEARS_PAST + 1) });
       const data = {
         ...mockData,
-        providerFacility: [{ treatmentDateRange: { to } }],
+        [EVIDENCE_VA]: true,
+        locations: [{ evidenceDates: { to } }],
       };
       const page = setup({ index: 0, data });
       const { container } = render(page);
@@ -645,7 +796,7 @@ describe('<EvidencePrivateRecords>', () => {
       await $('va-memorable-date').__events.dateBlur(toBlurEvent);
 
       waitFor(async () => {
-        const dateTo = await $$('va-memorable-date', container)?.[1];
+        const dateTo = await $$('va-memorable-date', container)[1];
         expect(dateTo.error).to.contain(errorMessages.evidence.newerDate);
         expect(dateTo.invalidMonth).to.be.false;
         expect(dateTo.invalidDay).to.be.false;
@@ -658,35 +809,36 @@ describe('<EvidencePrivateRecords>', () => {
       const to = parseDateWithOffset({ years: -10 });
       const data = {
         ...mockData,
-        providerFacility: [{ treatmentDateRange: { from, to } }],
+        [EVIDENCE_VA]: true,
+        locations: [{ evidenceDates: { from, to } }],
       };
       const page = setup({ index: 0, data });
       const { container } = render(page);
 
+      const dateTo = await $$('va-memorable-date', container)[1];
       // blur date to input
       await $('va-memorable-date').__events.dateBlur(toBlurEvent);
 
-      waitFor(async () => {
-        const dateTo = await $$('va-memorable-date', container)?.[1];
-        expect(dateTo.error).to.contain(sharedErrorMessages.endDateBeforeStart);
-        expect(dateTo.invalidMonth).to.be.true;
-        expect(dateTo.invalidDay).to.be.true;
-        expect(dateTo.invalidYear).to.be.true;
+      waitFor(() => {
+        expect(dateTo.error).to.contain(sharedErrorMessage.endDateBeforeStart);
       });
     });
 
     it('should show an error when the issue is not unique', async () => {
       const data = {
         ...mockData,
-        providerFacility: [mockFacility, mockFacility],
+        [EVIDENCE_VA]: true,
+        locations: [mockLocation, mockLocation],
       };
       const page = setup({ index: 1, data });
       const { container } = render(page);
 
-      const input = $('va-text-input', container);
+      const input = await $('va-text-input', container);
       await fireEvent.blur(input);
 
-      expect(input.error).to.contain(errorMessages.evidence.uniquePrivate);
+      waitFor(() => {
+        expect(input.error).to.contain(errorMessages.evidence.uniqueVA);
+      });
     });
 
     it('should show no contestable issues were selected message', () => {
@@ -695,6 +847,40 @@ describe('<EvidencePrivateRecords>', () => {
       expect($('va-checkbox-group', container).textContent).to.contain(
         NO_ISSUES_SELECTED,
       );
+    });
+
+    it('should not move focus after submitting & blurring first input', async () => {
+      const goSpy = sinon.spy();
+      const index = 1;
+      const data = {
+        ...mockData,
+        showScNewForm: true,
+        [EVIDENCE_VA]: true,
+        locations: [mockLocation, {}, mockLocation2],
+      };
+      const page = setup({
+        index,
+        goForward: goSpy,
+        goToPath: goSpy,
+        data,
+      });
+      const { container } = render(page);
+
+      // continue
+      clickContinue(container);
+
+      const input = await $('va-text-input', container);
+      await fireEvent.blur(input);
+      await userEvent.tab(); // ensure focus is not trapped; see #98137
+
+      waitFor(() => {
+        expect(getErrorElements(container).length).to.eq(3);
+        expect(goSpy.called).to.be.false;
+        expect(focusElementSpy.args[0][0]).to.eq('[role="alert"]');
+        expect(document.activeElement).to.not.eq(input);
+        // focus called one additional time when focus shifts back to the error
+        expect(focusElementSpy.args.length).to.be.lessThan(3);
+      });
     });
   });
 });
